Extract enum-cast compiler fixtures
